+++ src/Core/Visma.Yuki.Blog.Application/Ports/Driven/IPostPort.cs
@@ -1,0 +1,10 @@
+using Visma.Yuki.Blog.Domain.Entities;
+
+namespace Visma.Yuki.Blog.Application.Ports.Driven;
+
+public interface IPostPorts
+{
+    Task AddAsync(Post post, CancellationToken cancellationToken = default);
+    Task<IEnumerable<Post>> GetAllAsync(bool includeAuthor, CancellationToken cancellationToken);
+    Task<Post?> GetPostByIdAsync(Guid id, bool includeAuthor, CancellationToken cancellationToken);
+}