+++ src/Tests/Visma.Yuki.Blog.Tests.Integration/Endpoints/CreateAuthorEndpointTests.cs
@@ -111,7 +111,7 @@ public class CreateAuthorEndpointTests : IClassFixture<IntegrationTestWebAppFact
         await using var conn = new NpgsqlConnection(_connectionString);
         await conn.OpenAsync();
         await using var cmd = conn.CreateCommand();
-        cmd.CommandText = "TRUNCATE TABLE authors;";
+        cmd.CommandText = "TRUNCATE TABLE posts, authors CASCADE;";
         await cmd.ExecuteNonQueryAsync();
     }